Moving straggling campaign helpers into dedicated file.
