+++ resources/assets/components/CampaignPageNavigation/CampaignPageNavigation.js
@@ -1,9 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { isActionPage } from '../../helpers';
-import { prepareCampaignPageSlug } from '../../helpers/campaign';
 import PageNavigation from '../utilities/PageNavigation/PageNavigation';
+import { isActionPage, prepareCampaignPageSlug } from '../../helpers/campaign';
 import CampaignSignupFormContainer from '../CampaignSignupForm/CampaignSignupFormContainer';
 
 const CampaignPageNavigation = ({
