+++ resources/assets/components/ContentfulEntry/ContentfulEntryContainer.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 
 import ContentfulEntry from './ContentfulEntry';
-import { findContentfulEntry } from '../../helpers';
+import { findContentfulEntryInCampaign } from '../../helpers/campaign';
 
 /**
  * Provide state from the Redux store as props for this component.
@@ -9,7 +9,7 @@ import { findContentfulEntry } from '../../helpers';
  * @return {Object}
  */
 const mapStateToProps = (state, { id }) => ({
-  json: findContentfulEntry(state, id),
+  json: findContentfulEntryInCampaign(state, id),
 });
 
 export default connect(mapStateToProps)(ContentfulEntry);
