+++ resources/assets/helpers/campaign.js
@@ -1,24 +1,45 @@
 import { join } from 'path';
-import get from 'lodash/get';
+import { find, get } from 'lodash';
 import gql from 'graphql-tag';
+import { isBefore } from 'date-fns';
 
 import { getUserId } from './auth';
 
-/**
- * Prepare a campaign subpage's slug.
- *
- * @param  {String} campaignSlug
- * @param  {String} pageSlug
- * @return {String}
- */
-export function prepareCampaignPageSlug(campaignSlug, pageSlug) {
-  const missingCampaignSlug = pageSlug.indexOf(campaignSlug) < 0;
-
-  if (missingCampaignSlug) {
-    return join('/us/campaigns', campaignSlug, pageSlug);
+export const CAMPAIGN_SIGNUP_QUERY = gql`
+  query CampaignSignupQuery($userId: String!, $campaignId: String!) {
+    signups(userId: $userId, campaignId: $campaignId) {
+      id
+      group {
+        id
+        goal
+        name
+        groupTypeId
+        groupType {
+          name
+        }
+      }
+    }
   }
+`;
 
-  return join('/us/campaigns', pageSlug);
+/**
+ * Find an entry from within the campaign by given ID or Slug param.
+ * (Returns false if not found).
+ *
+ * @param  {Object} state
+ * @param  {String} identifier
+ * @return {Object|Undefined}
+ */
+export function findContentfulEntryInCampaign(state, identifier) {
+  const campaign = state.campaign;
+
+  const contentfulEntries = [].concat(campaign.pages, campaign.quizzes);
+
+  return find(
+    contentfulEntries,
+    entry =>
+      entry.id === identifier || get(entry, 'fields.slug') === identifier,
+  );
 }
 
 /**
@@ -54,24 +75,9 @@ export function getCampaignFaqPath() {
   return faqSlug ? `/us/campaigns/${faqSlug}` : undefined;
 }
 
-export const CAMPAIGN_SIGNUP_QUERY = gql`
-  query CampaignSignupQuery($userId: String!, $campaignId: String!) {
-    signups(userId: $userId, campaignId: $campaignId) {
-      id
-      group {
-        id
-        goal
-        name
-        groupTypeId
-        groupType {
-          name
-        }
-      }
-    }
-  }
-`;
-
 /**
+ * Get a query variables for a campaign signup.
+ *
  * @return {Object}
  */
 export function getCampaignSignupQueryVariables() {
@@ -79,6 +85,47 @@ export function getCampaignSignupQueryVariables() {
     userId: getUserId(),
     campaignId: getCampaign().campaignId,
   };
+}
+
+/**
+ * Determine if a page is an 'Action' page.
+ *
+ * @param  {Object} page
+ * @return {Boolean}
+ */
+export function isActionPage(page) {
+  return page.type === 'page' && page.fields.slug.endsWith('action');
+}
+
+/**
+ * Check if the given campaign is closed.
+ *
+ * @param  {String}  endDate
+ * @return {Boolean}
+ */
+export function isCampaignClosed(endDate) {
+  if (!endDate) {
+    return false;
+  }
+
+  return isBefore(endDate, new Date());
+}
+
+/**
+ * Prepare a campaign subpage's slug.
+ *
+ * @param  {String} campaignSlug
+ * @param  {String} pageSlug
+ * @return {String}
+ */
+export function prepareCampaignPageSlug(campaignSlug, pageSlug) {
+  const missingCampaignSlug = pageSlug.indexOf(campaignSlug) < 0;
+
+  if (missingCampaignSlug) {
+    return join('/us/campaigns', campaignSlug, pageSlug);
+  }
+
+  return join('/us/campaigns', pageSlug);
 }
 
 export default null;
